% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{myadd}
\alias{myadd}
\title{加法}
\usage{
myadd(x = 1, y = 2)
}
\arguments{
\item{x}{参数x}

\item{y}{参数y}
}
\value{
返回值
}
\description{
加法
}
\examples{
a =myadd()
b =myadd(3,4)
c =myadd(x=5,y=6)
d =myadd(y=2,x=3)
}
